(#4) Move all the code dealing with initializing the http connection pools into its own file.
